requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
bleach==6.2.0
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
django==5.2.5
django-appconf==1.1.0
django-autoslug==1.9.9
django-cleanup==9.0.0
django-filebrowser==4.0.3
django-grappelli==3.0.10
django-selectable==1.5.0
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
pyjwt==2.10.1
pyparsing==3.2.3
pytz==2025.2
requests==2.32.5
six==1.10.0
soupsieve==2.7
sqlparse==0.5.3
typing-extensions==4.14.1
urllib3==2.5.0
webencodings==0.5.1
